refactor: api 파일에서 query 훅 모델로 분리
