+++ src/entities/comment/api/commentApi.ts
@@ -1,17 +1,15 @@
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-
 import { Comment } from '../model/types';
 
 const API_BASE_URL = '/api/comments';
 
-const fetchCommentsByPostId = async (postId: number): Promise<Comment[]> => {
+export const fetchCommentsByPostId = async (postId: number): Promise<Comment[]> => {
   const response = await fetch(`${API_BASE_URL}/post/${postId}`);
   if (!response.ok) throw new Error('Failed to fetch comments');
   const data = await response.json();
   return data.comments;
 };
 
-const addComment = async (newComment: {
+export const addComment = async (newComment: {
   body: string;
   postId: number;
   userId: number;
@@ -25,7 +23,7 @@ const addComment = async (newComment: {
   return response.json();
 };
 
-const updateComment = async ({
+export const updateComment = async ({
   commentId,
   body,
 }: {
@@ -41,7 +39,7 @@ const updateComment = async ({
   return response.json();
 };
 
-const deleteComment = async (commentId: number): Promise<{ isDeleted: boolean }> => {
+export const deleteComment = async (commentId: number): Promise<{ isDeleted: boolean }> => {
   const response = await fetch(`${API_BASE_URL}/${commentId}`, {
     method: 'DELETE',
   });
@@ -49,7 +47,7 @@ const deleteComment = async (commentId: number): Promise<{ isDeleted: boolean }>
   return response.json();
 };
 
-const likeComment = async ({
+export const likeComment = async ({
   commentId,
   currentLikes,
 }: {
@@ -63,68 +61,4 @@ const likeComment = async ({
   });
   if (!response.ok) throw new Error('Failed to like comment');
   return response.json();
-};
-
-/**
- * @description 특정 게시물의 댓글 목록을 가져오는 useQuery 훅
- * @param postId 댓글을 가져올 게시물의 ID
- */
-export const useFetchComments = (postId: number | null) => {
-  return useQuery({
-    queryKey: ['comments', postId],
-    queryFn: () => fetchCommentsByPostId(postId!),
-    enabled: !!postId,
-  });
-};
-
-/**
- * @description 새 댓글을 추가하는 useMutation 훅
- */
-export const useAddComment = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: addComment,
-    onSuccess: (data) => {
-      queryClient.invalidateQueries({ queryKey: ['comments', data.postId] });
-    },
-  });
-};
-
-/**
- * @description 댓글을 수정하는 useMutation 훅
- */
-export const useUpdateComment = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: updateComment,
-    onSuccess: (data) => {
-      queryClient.invalidateQueries({ queryKey: ['comments', data.postId] });
-    },
-  });
-};
-
-/**
- * @description 댓글을 삭제하는 useMutation 훅
- */
-export const useDeleteComment = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: ({ commentId }: { commentId: number; postId: number }) => deleteComment(commentId),
-    onSuccess: (_, variables) => {
-      queryClient.invalidateQueries({ queryKey: ['comments', variables.postId] });
-    },
-  });
-};
-
-/**
- * @description 댓글에 '좋아요'를 추가하는 useMutation 훅
- */
-export const useLikeComment = () => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationFn: likeComment,
-    onSuccess: (data) => {
-      queryClient.invalidateQueries({ queryKey: ['comments', data.postId] });
-    },
-  });
 };
